@@ -1,9 +1,80 @@
+"""rkstiff.etd
+Provides classes and functions for Exponential Time Differencing Runge-Kutta methods
+"""
+
 import numpy as np
-from rkstiff.solver import StiffSolverAS, StiffSolverCS
+from rkstiff.solver import StiffSolverAS, StiffSolverCS, SolverConfig
+
+
+class ETDConfig:
+    """
+    Configuration parameters for the ETD-based solvers.
+
+    Attributes
+    ----------
+    modecutoff : float
+        Numerical inaccuracy cutoff for psi functions (default 0.01).
+        Must be between 0.0 and 1.0.
+    contour_points : int
+        Number of points for the contour integral (default 32).
+        Must be an integer greater than 1.
+    contour_radius : float
+        Radius of the circular contour integral (default 1.0).
+        Must be greater than 0.
+    """
+
+    def __init__(self, modecutoff=0.01, contour_points=32, contour_radius=1.0):
+        """Initialize ETDConfig with validated parameters."""
+        self._modecutoff = None
+        self._contour_points = None
+        self._contour_radius = None
+
+        # Use setters for validation
+        self.modecutoff = modecutoff
+        self.contour_points = contour_points
+        self.contour_radius = contour_radius
+
+    @property
+    def modecutoff(self):
+        """Get the modecutoff value."""
+        return self._modecutoff
+
+    @modecutoff.setter
+    def modecutoff(self, value):
+        """Set the modecutoff value with validation."""
+        if (value > 1.0) or (value <= 0):
+            raise ValueError(f"modecutoff must be between 0.0 and 1.0 but is {value}")
+        self._modecutoff = value
+
+    @property
+    def contour_points(self):
+        """Get the contour_points value."""
+        return self._contour_points
+
+    @contour_points.setter
+    def contour_points(self, value):
+        """Set the contour_points value with validation."""
+        if not isinstance(value, int):
+            raise TypeError(f"contour_points must be an integer but is {value}")
+        if value <= 1:
+            raise ValueError(f"contour_points must be an integer greater than 1 but is {value}")
+        self._contour_points = value
+
+    @property
+    def contour_radius(self):
+        """Get the contour_radius value."""
+        return self._contour_radius
+
+    @contour_radius.setter
+    def contour_radius(self, value):
+        """Set the contour_radius value with validation."""
+        if value <= 0:
+            raise ValueError(f"contour_radius must greater than 0 but is {value}")
+        self._contour_radius = value
 
 
 def phi1(z):
-    """Computes RKETD psi-function of the first order.
+    """compute_s RKETD psi-function of the first order.
     INPUTS
         z - real or complex-valued input array
     OUTPUT
@@ -13,7 +84,7 @@ def phi1(z):
 
 
 def phi2(z):
-    """Computes RKETD psi-function of the second order.
+    """compute_s RKETD psi-function of the second order.
     INPUTS
         z - real or complex-valued input array
     OUTPUT
@@ -23,7 +94,7 @@ def phi2(z):
 
 
 def phi3(z):
-    """Computes RKETD psi-function of the third order.
+    """compute_s RKETD psi-function of the third order.
     INPUTS
         z - real or complex-valued input array
     OUTPUT
@@ -50,58 +121,9 @@ class ETDAS(StiffSolverAS):
 
     """
 
-    def __init__(
-        self,
-        linop,
-        NLfunc,
-        epsilon=1e-4,
-        incrF=1.25,
-        decrF=0.85,
-        safetyF=0.8,
-        adapt_cutoff=0.01,
-        minh=1e-16,
-        modecutoff=0.01,
-        contour_points=32,
-        contour_radius=1.0,
-    ):
-        super().__init__(
-            linop,
-            NLfunc,
-            epsilon=epsilon,
-            incrF=incrF,
-            decrF=decrF,
-            safetyF=safetyF,
-            adapt_cutoff=adapt_cutoff,
-            minh=minh,
-        )
-        self.modecutoff = modecutoff
-        if (self.modecutoff > 1.0) or (self.modecutoff <= 0):
-            raise ValueError(
-                "modecutoff must be between 0.0 and 1.0 but is {}".format(
-                    self.modecutoff
-                )
-            )
-        self.contour_points = contour_points
-        if not isinstance(self.contour_points, int):
-            raise TypeError(
-                "contour_points must be an integer but is {}".format(
-                    self.contour_points
-                )
-            )
-        if self.contour_points <= 1:
-            raise ValueError(
-                "contour_points must be an integer greater than 1 but is {}".format(
-                    self.contour_points
-                )
-            )
-
-        self.contour_radius = contour_radius
-        if self.contour_radius <= 0:
-            raise ValueError(
-                "contour_radius must greater than 0 but is {}".format(
-                    self.contour_radius
-                )
-            )
+    def __init__(self, lin_op, nl_func, config=SolverConfig(), etd_config=ETDConfig()):
+        super().__init__(lin_op, nl_func, config)
+        self.etd_config = etd_config
         self._h_coeff = None
 
 
@@ -123,35 +145,7 @@ class ETDCS(StiffSolverCS):
 
     """
 
-    def __init__(
-        self, linop, NLfunc, modecutoff=0.01, contour_points=32, contour_radius=1.0
-    ):
-        super().__init__(linop, NLfunc)
-        self.modecutoff = modecutoff
-        if (self.modecutoff > 1.0) or (self.modecutoff <= 0):
-            raise ValueError(
-                "modecutoff must be between 0.0 and 1.0 but is {}".format(
-                    self.modecutoff
-                )
-            )
-        self.contour_points = contour_points
-        if not isinstance(self.contour_points, int):
-            raise TypeError(
-                "contour_points must be an integer but is {}".format(
-                    self.contour_points
-                )
-            )
-        if self.contour_points <= 1:
-            raise ValueError(
-                "contour_points must be an integer greater than 1 but is {}".format(
-                    self.contour_points
-                )
-            )
-        self.contour_radius = contour_radius
-        if self.contour_radius <= 0:
-            raise ValueError(
-                "contour_radius must greater than 0 but is {}".format(
-                    self.contour_radius
-                )
-            )
+    def __init__(self, lin_op, nl_func, etd_config=ETDConfig()):
+        super().__init__(lin_op, nl_func)
+        self.etd_config = etd_config
         self._h_coeff = None
